+++ frontend/src/routes/router.tsx
@@ -1,11 +1,9 @@
 import { Outlet, createBrowserRouter } from 'react-router-dom';
 
-import Header from '@/components/Header/Header';
 import Template from '@/pages/Template';
 import TemplateList from '@/pages/TemplateList';
 import UploadsTemplate from '@/pages/UploadsTemplate';
-
-
+import { Header } from '@/components/Header';
 
 const Layout = () => {
   const style = { maxWidth: '1024px', margin: 'auto', padding: '0 2rem' };
